Reorganize tests (#819)
