+++ twindle-cli/tests/twitter/unit/api/twitter-endpoints/tweets.unit.test.js
@@ -1,5 +1,5 @@
-const { ApiErrors } = require("../../error");
-const { getTweetById } = require("./tweets");
+const { ApiErrors } = require("../../../../../src/twitter/error");
+const { getTweetById } = require("../../../../../src/twitter/api/twitter-endpoints/tweets");
 
 describe("twitter |> api |> twitter-endpoints |> tweets module", () => {
   it("should throw error if tweet id is falsey", () => {
